Renamed: component PickerSearchQuery.vue
