+++ apps/frontend/src/components/page-products/PickerSearchQuery.vue
